+++ main_app/receive_sms.py
@@ -1,13 +1,28 @@
 #!/util/python3/bin/python
+from enum import Enum
+from io import BytesIO
+
+import requests
+from django.conf import settings
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
+from loguru import logger
+from PIL import Image
+from pydantic import BaseModel
+from rest_framework.status import HTTP_200_OK
 
 # from django_twilio.decorators import twilio_view
 from twilio.twiml.messaging_response import MessagingResponse
 
 from main_app import contribution_database as database
-from main_app import crowdhydrology_website_database as website_database
 from main_app.models import Station
+from model.detection import ContributionImageDetector, GeminiClient
+from model.exceptions import (
+    INVALID_GAUGE_READING_EXCEPTION,
+    INVALID_STATION_LABEL_EXCEPTION,
+    InvalidBoxesException,
+)
+from model.preprocessor import GaugePreprocessor, StationLabelPreprocessor
 
 """
 Functions to receive and parse sms.
@@ -20,20 +35,152 @@ Created: 06/18/2018
 """
 
 
+class AcceptedMediaTypes(Enum):
+    jpeg = "image/jpeg"
+    jpg = "image/jpg"
+    png = "image/png"
+    # heic = "image/heic"
+
+
+class IncomingMMS(BaseModel):
+    """Model for incoming MMS messages."""
+
+    media_url: str
+    media_type: AcceptedMediaTypes
+
+
+class TwilioMediaException(Exception):
+    """Custom exception for HTTP errors."""
+
+    def __init__(self):
+        super.__init__("Failed to retrieve the media. Please try again.")
+
+
+PROMPT_TEXT = """
+    Task: You are given two images in a single prompt.
+
+    Image 1: Staff Gauge
+
+    - Decide if this is a clear staff-gauge photo.
+    - If it’s not a staff gauge, or if it’s too unclear for a confident reading (confidence < 0.70),
+    - set "is_valid_gauge": false and stop.
+    - Otherwise, calculate the exact water-level reading at the red line.
+    - The gauge reading is always a positive floating-point number in 2 decimal places.
+
+    Gauge Details:
+    - Major stripes: longer, labeled marks (e.g. 1.0, 1.1, …).
+    - Minor stripes: shorter, evenly spaced between two majors.
+
+    Step-by-Step Instructions:
+    - Detect two consecutive, fully visible major stripes and note their labels (e.g. 1.0 & 1.1).
+    - Count the minor stripes between them; compute
+    minor_unit = (major2_label − major1_label) ÷ minor_count_between.
+    - Locate the red waterline.
+    - Identify the first major stripe above that line; record its label M.
+    - Count how many minor stripes lie between the waterline and stripe M; call that n.
+    - Compute reading = M + (n × minor_unit).
+
+    Image 2: Station Label
+    - Analyze the image to verify if it is a valid station label.
+    - A valid station label contains a station ID that matches one of the predefined station IDs.
+    - If not valid, respond with "is_valid_station_label": false and "station_id": null.
+    - If valid, set "is_valid_station_label": true and return the "station_id".
+
+    Critical Consideration:
+    - If the image is beyond the ability to analyze, unreadable,
+    or if the confidence of the output is below 40%, mark it invalid.
+"""
+
+CONTRIBUTION_EXCEPTION_MESSAGE = (
+    "An error occurred while processing your contribution. Please try again later."
+)
+
+
 @csrf_exempt
 # @twilio_view  # Visit link for more info https://www.twilio.com/blog/2014/04/building-a-simple-sms-message-application-with-twilio-and-django-2.html
 # NOTE: twilio_view decorator is deprecated due to no support for Django > 5.0
 # TODO: Implement RequestValidator from twilio.request_validator
 # TODO: Sample implementation for RequestValidator: @twilio_view impl -> https://github.com/rdegges/django-twilio/blob/master/django_twilio/decorators.py
 def incoming_sms(request):
+    # Start our TwiML response
+    resp = MessagingResponse()
+
+    num_media = int(request.POST.get("NumMedia", 0))
+    if num_media == 1:  # if media received.
+        try:
+            # Handle incoming MMS with one media item
+            mms = IncomingMMS(
+                media_url=request.POST.get("MediaUrl0"),
+                media_type=request.POST.get("MediaContentType0"),
+            )
+            # Get gauge measurement.
+            slp, gp = StationLabelPreprocessor(), GaugePreprocessor()
+            detector = ContributionImageDetector(
+                f"{settings.BASE_DIR}/model/models/best.pt"
+            )
+
+            media = requests.get(mms.media_url)
+            if media.status_code != HTTP_200_OK:
+                raise TwilioMediaException()
+
+            media = Image.open(BytesIO(media.content))
+
+            detected = detector.detect(media)  # Detect the ROIs
+
+            # Extract ROIs.
+            station_label_roi = detector.get_station_label_roi(detected[0])
+            gauge_roi = detector.get_gauge_roi(detected[0])
+
+            station_label_roi = slp.preprocess(station_label_roi)
+            gauge_roi = gp.preprocess(gauge_roi)
+
+            # Extract reading values.
+            llm_client = GeminiClient(secret_key="")
+
+            logger.warning(
+                "Extracting gauge and station label reading from the image..."
+            )
+            contribution = llm_client.get_gauge_and_station_label_reading(
+                PROMPT_TEXT, gauge_roi, station_label_roi
+            )
+            logger.success(
+                "Successfully extracted gauge and station label reading from the image."
+            )
+            if not contribution.station_label.is_valid_station_label:
+                raise InvalidBoxesException(INVALID_STATION_LABEL_EXCEPTION)
+            if not contribution.gauge_reading.is_valid_gauge:
+                raise InvalidBoxesException(INVALID_GAUGE_READING_EXCEPTION)
+
+            # Save Contribution
+
+        except InvalidBoxesException as e:  # Image not visible
+            resp.message(
+                "It seems that the image is not clear or is invalid. Please try again."
+            )
+            logger.error(f"Error: {e.message}, ")
+            return HttpResponse(str(resp), content_type="application/xml")
+
+        except ValueError:
+            resp.message(
+                "The media type is not supported. Please send a JPEG, JPG, or PNG image."
+            )
+            return HttpResponse(str(resp), content_type="application/xml")
+        except TwilioMediaException as e:
+            resp.message(str(e))
+            return HttpResponse(str(resp), content_type="application/xml")
+        except Exception as e:
+            logger.error(e)
+            resp.message(CONTRIBUTION_EXCEPTION_MESSAGE)
+            return HttpResponse(str(resp), content_type="application/xml")
+
+        return HttpResponse(str(resp), content_type="application/xml")
+
+    """Handle from text message."""
     # Get the text message the user sent to our Twilio number
     message_body = request.POST.get("Body", None)
     phone_number = request.POST.get("From", None)
 
     is_valid, station_id, water_height, temperature, error_msg = parse_sms(message_body)
-
-    # Start our TwiML response
-    resp = MessagingResponse()
 
     if is_valid:
         reply_msg = "Thanks for contributing to CrowdHydrology research and being a citizen scientist!"
@@ -74,8 +221,8 @@ def incoming_sms(request):
     database.save_contribution(
         is_valid, station_id, water_height, temperature, phone_number, message_body
     )
-    if is_valid:
-        website_database.save_contributions_to_csv(station_id)
+    # if is_valid:
+    #     website_database.save_contributions_to_csv(station_id)
 
     # TODO: send sms asynchronously.
     return HttpResponse(str(resp), content_type="application/xml")
