+++ model/exceptions.py
@@ -1,3 +1,11 @@
+INVALID_STATION_LABEL_EXCEPTION = (
+    "The station label is invalid or not recognized. Please try again."
+)
+INVALID_GAUGE_READING_EXCEPTION = (
+    "The gauge reading is invalid or not recognized. Please try again."
+)
+
+
 class InvalidBoxesException(Exception):
     """Exception raised for invalid number of bounding boxes detected."""
